+++ packages/sync/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.48.0 (2026-06-10)
+
 -   Prevent RTC polling interval filters from slowing active HTTP polling.
 
 ## 1.47.0 (2026-05-27)